+++ doc/changelog.md
@@ -16,6 +16,7 @@
 - Added `WUFFS_BASE__PIXEL_FORMAT__BGR_565`.
 - Added interfaces.
 - Added preprocessor.
+- Added single-quoted strings.
 - Added tokens.
 - Changed `gif.decoder_workbuf_len_max_incl_worst_case` from 1 to 0.
 - Made `wuffs_base__pixel_format` a struct.